refactor(query-param): move to separate file
